저장, 수정을 같은 메서드에서 처리하기로 수정

+++ src/main/java/site/dolshei/emspring/mvc/controller/BoardController.java
@@ -38,11 +38,12 @@ public class BoardController {
 
     /**
      * 게시글 저장 / 수정
-     * @param board
+     * @param parameter
      */
     @GetMapping("/save")
-    public int save(Board board) {
-        return boardService.save(board);
+    public int save(Board parameter) {
+        boardService.save(parameter);
+        return parameter.getBoardSeq();
     }
 
     /**
@@ -50,8 +51,13 @@ public class BoardController {
      * @param boardSeq
      */
     @GetMapping("/delete/{boardSeq}")
-    public void delete(@PathVariable int boardSeq) {
+    public boolean delete(@PathVariable int boardSeq) {
+        Board board = boardService.get(boardSeq);
+        if (board == null) {
+            return false;
+        }
         boardService.delete(boardSeq);
+        return true;
     }
 
 }
